[PureFluid] Use pressure criteria to decide phase
Use the pressure compared with the saturation pressure to decide gas or
liquid phase, rather than temperature. This change is required because
for some substances, the saturation temperature for low values of the
pressure was outside the limits of the equation of state, resulting
in convergence errors in the saturation solver.

Since phaseOfMatter can throw, we need to translate exceptions in
the Cython interface as well.

Add tests to prevent regression with cases identifed in #786.
Fixes #786.