[Test] Fix test failures when re-running tests
The liquidvapor.yaml generated by the test suite from the XML or CTI input
files is not the same as the one in the Cantera data directory because new
names for some of the phase names have been added (such as 'carbon-dioxide'
instead of 'carbondioxide') which are not in the old input files.

Testing errors would occur when the generated liquidvapor.yaml was then read
instead of the version from the Cantera data directory, which would happen
when re-running the test suite, since normally the liquidvapor tests run
before the conversion tests.