[Input] Make accessing input file data outside phase node easier
Some phase types utilize information in the input file that is
contained outside the phase definition (e.g. phase definitions
for nested phases). This change to ThermoPhase::setParameters
makes it easier for those phases to have access to the full input
file definition without re-reading and reparsing the YAML file.