[Examples] Properly close multiprocessing.Pool
The multiprocess.Pool object must be closed properly, or the process can
hang according to the Python docs. The recommended method for short code
blocks is to use the context manager. The format of the arguments to
Pool is chosen to allow the last line to be de-dented to remove
ambiguity about the definition of the context manager and the content.