[purefluid] Make pure saturated liquid/vapor part of TwoPhase region
Saturated liquid and saturated vapor form the boundary of the vapor dome and
are included in the phase-of-matter labeled 'liquid-vapor-mix'. At the same
time, properties that are based on finite difference calculations of a pure
liquid or pure vapor can be calculated. This approach corresponds to the
implementation of the CoolProp package (http://www.coolprop.org).