[1D] Fix error when compiling with GLIBCXX_ASSERTIONS enabled
If the rightmost domain has no solution components of its own, calling
functions which need a pointer to the "local" portion of the solution
vector by taking &m_x[domain(n).loc()] gives a pointer just past the
end of m_x, but this pointer is not used in the function called.

Adding a special case for this to OneDim::start avoids having this be
reported as an error when using bounds checking features like those
enabled by the _GLIBCXX_ASSERTIONS preprocessor macro.

Partially resolves #901