[Sphinx] Fix spurious warnings in __init__ docs
This has been bugging me for years, and I'm very happy to have finally
fixed it! To debug this, I used the -vv output from sphinx-build to
inspect what autodoc was doing. autodoc seemed to have no problem
reading the source, but it was indeed generating reST with unmatched
emphasis and strong markers. Although it really seems like autodoc
should be able to handle this case, apparently it can't. Then, I was
reading the autodoc documentation when I noticed this idea of an event
handler, a function that gets called whenever a particular event
happens. I realized I could get in and munge the docstring text to
replace any unescaped splats with escaped ones.