[CI] Speed up coverage build
Compiling with the NDEBUG macro defined skips assertion checks, such
as those in standard library code. Skipping these checks allows the
tests to run slightly faster.

Compiling with a system copy of Sundials speeds up running the test
suite slightly because the Sundials library being used is still
compiled with optimizations turned on, as opposed to the code within
Cantera, that has to be compiled without optimizations so that the
line coverage will be accurate.