[SCons/Samples] Fix finding OMP library
It is possible for the OMP header to be installed in the search
path, but the library not to be available. This can happen, for
example on macOS when Boost and GCC are installed by Homebrew,
the Homebrew include path is set to find Boost, and the Apple
compilers are being used.