[Input] Fix YAML file-based constructor for SurfPhase
Species are added before phase-specific parameters like site density
are set. Adding the species sets their coverages, which uses the site
density. Therefore, the initial value for the site density needs to be
non-zero.

This only affected the specific SurfPhase constructor, not methods
which use ThermoFactory, since in that case the default constructor
for SurfPhase gets called, which does set a non-zero site density.