[Test] Replace square wave with triangle wave in test_mass_flow_controller
Use a triangular pulse in the mass flow rate instead of a square wave
so that the function is continuous (but not necessarily smooth). This
still shows that CVODES is able to appropriately reduce the step size
around such a feature in the input function, but seems to be give
results that are more consistent across different systems / library
versions.

Fixes #1033.