[SCons] Make Fortran compiler and options consistent
In some cases, the setting of "SHF90" (the compiler to use for
compiling shared objects from F90 code) does not inherit the setting
of "F90" (the compiler to use for compiling normal objects from F90
code). This changes these two to be consistent, as well as the
corresponding variables for other Fortran compilers.