[Cython] Use setup.cfg to configure the package
This change mirrors the one for the source distribution. This change
results in a non-executable template file
(setup.cfg.in), which makes edits easier.

This change
removes the requirement to template the Python extension into setup.py,
and adds a Setuptools extension module. This results in correct builds
when bdist_wheel is specified and uses package_data properly.