[Cython] Fix Windows builds
For some versions of Python, the DLL module extension included in the
sysconfig module is not correct. In that case, the file extension needs
to be constructed manually.

To make this information easier to get to, the JSON module is used to
parse data returned by sysconfig from the Python which will be building
the module.