Implement adaptive preconditioning for reactor networks
Adds PreconditionerBase and derives AdaptivePreconditioner from this
base class. PreconditionerBase is intended to be an abstract class so
other types of preconditioners can be extended. AdaptivePreconditioner
is a specific type of preconditioning based on a tolerance.

The preconditioned integrator works with all Sundials versions supported
by Cantera.

The ILUT algorithm is used to limit fill-in when factorizing the preconditioner.

Adds access to CVODES integrator statistics in C++ and Python

Adds reactor classes "IdealGasConstPressureMoleReactor" and "IdealGasMoleReactor"
that use a mole-based state vector to increase the sparsity of the preconditioner.