[1D] Fix error when reporting singular Jacobian
Fixed an error in the error handler for singular Jacobians where the
wrong domain was being identified as containing the problematic row,
that ultimately resulted in an integer divide by zero error instead of
a CanteraError identifying which domain, grid point, and solution
component the row corresponded to.

This fixes a regression introduced in 796cdcbf0.