[Python] Disable FlameBase.from_pandas
FlameBase.from_pandas is a vestigial method that is left over from an earlier
HDF interface that built on pandas and was replaced by the current
implementation between 2.4 and 2.5.1 (PR #840); the original HDF implementation
was never released.

There is some merit to allowing import, but further discussion is postponed
until after the release of Cantera 2.6 (see #1236).