[ck2yaml] Add support for format with single intermediate temperature
This format, used in the original NASA thermo database by McBride et al
stores the molecular weight in the last section of the first line of
each thermo entry.

Also, correct handling of cases where the inferred intermediate temperature
is outside the range (Tmin, Tmax) and the species only has a single set
of non-zero coefficients.