[python/examples] Examples of saving and loading flame solutions.
Examples of different ways to do this, saving via yaml, HDF5 and CSV,
then loading either as a restart or as an initial guess, including
with some modifications to the data (eg. fewer grid points, or species).
There are some subtleties, like you have to create the correct boundary
conditions before you load the initial guess.

There is one method (loading an initial guess from an HDF file) that
is currently not working, so this is anticipated.

Because of issue #1288 in 2.6.0 the required version is set to 3.0