[Python] Remove broken set_initial_guess(data='file.h5')
The ability to get an initial guess from an h5 file does not work.
Removing the broken attempt, documentation, and example that suggests it should.
As per ishoegl's suggestion https://github.com/Cantera/cantera/pull/1293#discussion_r886894943

For a work-around, restore a flame from the hdf5 file with 
flame.read_hdf then make a SolutionArray (or pandas.DataFrame) from that
and use one of those to create an initial guess (both of which work)