Implements MoleReactor and ConstPressureMoleReactor, tests, and docs.
This commit implements MoleReactor and ConstPressureMoleReactor classes,
it adds the appropriate python interfaces for them, it also adds a test
comparing surface chemistry of mole reactors to traditional reactors.

test_component_names had a bug because the network was not initialized
so self.net.n_vars was 0 and the loop was never entered. Adding a line
to initialize the network reveal that the test failed for
IdealGasMoleReactor so the appropriate lines were added to fix it.

Add wrappers for Extensible...MoleReactor, update docs, comment fixes.