Allow float input of integer geometry flags in ck2yaml
If `permissive` is given, `float` geometry flags are cast to `int` and are accepted if the values are equal, otherwise, `InputError` is raised. If `permissive` is not given, existing behavior is unchanged.

Add tests to check for correct behavior with and without permissive, for float rounding error, and for character inputs.

Closes #1395